Separated tests that require UI to run; added surefire plugin; made it so the github workflow does not run UI tests
