+++ .github/workflows/maven.yml
@@ -28,4 +28,4 @@ jobs:
         distribution: 'temurin'
         cache: maven
     - name: Build with Maven
-      run: mvn -B package --file pom.xml
+      run: mvn -B package --file pom.xml -Dsurefire.excludesFile=uiTestsExclude.txt
