+++ src/test/java/carleton/sysc4907/ui/AddCommandTest.java
@@ -1,8 +1,9 @@
-package carleton.sysc4907.command;
-
+package carleton.sysc4907.ui;
 import carleton.sysc4907.EditingAreaProvider;
+import carleton.sysc4907.command.AddCommand;
+import carleton.sysc4907.command.args.AddCommandArgs;
 import carleton.sysc4907.command.args.RemoveCommandArgs;
-import carleton.sysc4907.processing.ElementIdManager;
+import carleton.sysc4907.processing.ElementCreator;
 import carleton.sysc4907.view.DiagramElement;
 import org.junit.jupiter.api.Test;
 
@@ -30,8 +31,7 @@ import static org.mockito.Mockito.*;
 
 @ExtendWith(MockitoExtension.class)
 @ExtendWith(ApplicationExtension.class)
-public class RemoveCommandTest {
-
+public class AddCommandTest {
     @Mock
     private DiagramModel mockDiagramModel;
 
@@ -48,57 +48,46 @@ public class RemoveCommandTest {
     private Pane mockEditingArea;
 
     @Mock
-    private DiagramElement mockDiagramElement1;
+    private ElementCreator mockElementCreator;
 
     @Mock
-    private DiagramElement mockDiagramElement2;
-
-    @Mock
-    private ElementIdManager mockElementIdManager;
+    private DiagramElement mockDiagramElement;
 
     @Test
-    void executeRemoves() {
-    /* test that these are called:
-        Pane editingArea = EditingAreaProvider.getEditingArea();
-        editingArea.getChildren().removeAll(args.elements());
-        diagramModel.getElements().removeAll(args.elements());
-        diagramModel.getSelectedElements().clear();
-    */
+    void executeAdds() {
+        /* test that these are called:
+        editingArea.getChildren().add(obj);
+        diagramModel.getElements().add(element);
+         */
         //setup
-        long id1 = 1L;
-        long id2 = 2L;
         try (MockedStatic<EditingAreaProvider> utilities = Mockito.mockStatic(EditingAreaProvider.class)) {
             utilities.when(EditingAreaProvider::getEditingArea).thenReturn(mockEditingArea);
-            Mockito.when(mockDiagramElement1.getElementId()).thenReturn(id1);
-            Mockito.when(mockDiagramElement2.getElementId()).thenReturn(id2);
-            Mockito.when(mockElementIdManager.getElementById(id1)).thenReturn(mockDiagramElement1);
-            Mockito.when(mockElementIdManager.getElementById(id2)).thenReturn(mockDiagramElement2);
             Mockito.when(mockDiagramModel.getElements())
                     .thenReturn(mockElementsList);
-            Mockito.when(mockDiagramModel.getSelectedElements())
-                    .thenReturn(mockSelectedElementsList);
             Mockito.when(mockEditingArea.getChildren())
                     .thenReturn(mockNodesList);
-            Mockito.when(mockNodesList.removeAll(any(List.class))).
+            Mockito.when(mockEditingArea.getChildren())
+                    .thenReturn(mockNodesList);
+            Mockito.when(mockNodesList.add(any(Node.class))).
                     thenReturn(true);
-            Mockito.when(mockElementsList.removeAll(any(List.class)))
+            Mockito.when(mockElementsList.add(any(DiagramElement.class)))
                     .thenReturn(true);
-            doNothing().when(mockSelectedElementsList).clear();
-            List<DiagramElement> elemList = new ArrayList<>();
-            elemList.add(mockDiagramElement1);
-            elemList.add(mockDiagramElement2);
-            long[] idList = new long[] {mockDiagramElement1.getElementId(), mockDiagramElement2.getElementId()};
+            Mockito.when(mockElementCreator.create(eq("testType"), anyLong()))
+                    .thenReturn(mockDiagramElement);
 
-            RemoveCommandArgs args = new RemoveCommandArgs(idList);
-            RemoveCommand command = new RemoveCommand(args, mockDiagramModel, mockElementIdManager);
+            AddCommandArgs args = new AddCommandArgs("testType", 2L);
+            AddCommand command = new AddCommand(args, mockDiagramModel, mockElementCreator);
 
             //execute
             command.execute();
 
             //verify
-            verify(mockNodesList).removeAll(elemList);
-            verify(mockElementsList).removeAll(elemList);
-            verify(mockSelectedElementsList).clear();
+            verify(mockNodesList).add(any(Node.class));
+            verify(mockElementsList).add(any(DiagramElement.class));
         }
+
     }
+
+
+
 }
