+++ src/test/java/carleton/sysc4907/ui/view/DirectConnectionDialogTest.java
@@ -1,4 +1,4 @@
-package carleton.sysc4907.view;
+package carleton.sysc4907.ui.view;
 
 import carleton.sysc4907.DependencyInjector;
 import carleton.sysc4907.DiagramEditorLoader;
@@ -17,6 +17,7 @@ import org.testfx.api.FxRobot;
 import org.testfx.framework.junit5.ApplicationExtension;
 import org.testfx.framework.junit5.Start;
 import java.io.IOException;
+import java.util.concurrent.TimeUnit;
 
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
@@ -61,7 +62,7 @@ public class DirectConnectionDialogTest {
     }
 
     @Test
-    public void joinButtonClickedWithUsername(FxRobot robot) throws IOException {
+    public void joinButtonClickedWithUsername(FxRobot robot) throws IOException, InterruptedException {
         Mockito.doNothing().when(mockLoader).load(Mockito.any(Stage.class), Mockito.any(String.class), Mockito.any(String.class));
         robot.clickOn("#usernameField");
         robot.type(KeyCode.T, KeyCode.E, KeyCode.S, KeyCode.T);
@@ -74,6 +75,7 @@ public class DirectConnectionDialogTest {
         robot.press(KeyCode.SHIFT).press(KeyCode.SEMICOLON).release(KeyCode.SEMICOLON).release(KeyCode.SHIFT);
         robot.type(KeyCode.DIGIT1);
 
+        TimeUnit.MILLISECONDS.sleep(500);
         robot.clickOn(".button");
         Mockito.verify(mockLoader).load(Mockito.any(Stage.class), Mockito.any(String.class), Mockito.any(String.class));
     }
